requirements.txt
flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
redis==7.0.1
aiohttp==3.9.5
botbuilder-core==4.17.0
python-dotenv
ibm-watsonx-ai>=1.0.0
ibm-cloud-sdk-core>=3.16.0
